Integrated into adaptive_task and fixed mistake

+++ include/litmus.h
@@ -75,7 +75,7 @@ int set_rt_task_param(pid_t pid, struct rt_task* param);
  * @param r_work the relative work of this service level
  * @param qos the Quality Of Service from operating at this service level
  */
-void set_service_level_param(struct rt_task* tp, int s_level, int r_work, double qos);
+void set_service_level_param(struct rt_task* tp, int s_level, double r_work, double qos);
 
 /**
  * Get real-time task parameters for given process
